Refactor RustSec advisory loading into a rustsec.ts utils module
Move the advisory types and version-range parsing from version-ranges.ts,
together with the fetching/filtering/enrichment logic previously inlined in
the crate security page, into a shared $lib/utils/rustsec.ts module split
into fetchAdvisories (raw fetch) and loadAdvisories (filter + enrich).
